Refactor runtime loader component boundaries
Move checkpoint parsing into src.loader, MoE primitives into src.components.moe, and model-specific DeepSeek/MiniMax orchestration under src.models. Remove old runtime/model MoE namespaces without compatibility shims and update imports, package metadata, scripts, and namespace tests.

Validation: conda run -n deepseek bash -lc 'PYTHONPATH=$PWD python -m pytest tests/test_deepseek_v4_model_namespace.py -q' => 8 passed

Validation: conda run -n deepseek bash -lc 'PYTHONPATH=$PWD python -m pytest tests/test_partition_policy.py tests/test_moe_model_registry.py tests/test_minimax_m2_spec.py tests/test_minimax_m2_moe_runtime.py tests/test_gguf_bundle.py tests/test_gguf_iq1m_reader.py tests/test_gguf_imatrix.py tests/test_inspect_gguf_specs.py -q' => 48 passed, 1 skipped

Validation: conda run -n deepseek bash -lc 'PYTHONPATH=$PWD python -m compileall -q src tests'
