+++ setup.py
@@ -25,15 +25,16 @@ setup(
         "src",
         "src.cli",
         "src.encoding",
-        "src.gguf",
         "src.kernels",
+        "src.loader",
+        "src.loader.gguf",
+        "src.loader.mappings",
+        "src.components",
+        "src.components.moe",
         "src.models",
         "src.models.deepseek_v4",
         "src.models.minimax_m2",
-        "src.models.moe",
         "src.runtime",
-        "src.runtime.deepseek_v4",
-        "src.runtime.moe",
         "src.server",
     ],
     ext_modules=[
