+++ src/loader/__init__.py
@@ -1,0 +1,1 @@
+"""Checkpoint loading utilities and model weight-name mappings."""
