+++ src/models/minimax_m2/spec.py
@@ -2,10 +2,11 @@ from __future__ import annotations
 
 from collections import Counter
 
-from src.gguf.bundle import GGUFBundle
-from src.models.moe.capability import capability_status_for_role
-from src.models.moe.placement import HardwareProfile, heterogeneous_expert_decision, lowbit_device_resident_decision
-from src.models.moe.spec import (
+from src.loader.gguf.bundle import GGUFBundle
+from src.components.moe.capability import capability_status_for_role
+from src.components.moe.placement import HardwareProfile, heterogeneous_expert_decision, lowbit_device_resident_decision
+from src.loader.mappings.minimax_m2 import GLOBAL_TENSORS, LAYER_TENSORS, build_tensor_mappings, classify_tensor_name
+from src.components.moe.spec import (
     CapabilityItem,
     CapabilityReport,
     MoEArchitectureParams,
@@ -23,27 +24,8 @@ class MiniMaxM2Spec:
     architecture = "minimax-m2"
     display_name = "MiniMax-M2"
 
-    _GLOBAL_TENSORS = {
-        "token_embd.weight": ("embed_tokens.weight", "embedding", "transpose", "q4_k"),
-        "output.weight": ("lm_head.weight", "lm_head", "transpose", "q4_k"),
-        "output_norm.weight": ("final_norm.weight", "final_norm", "direct", "f32"),
-    }
-
-    _LAYER_TENSORS = {
-        "attn_q.weight": ("self_attn.q_proj.weight", "attn_q", "transpose", "q5_k"),
-        "attn_q_norm.weight": ("self_attn.q_norm.weight", "attn_norm", "direct", "f32"),
-        "attn_k.weight": ("self_attn.k_proj.weight", "attn_k", "transpose", "q5_k"),
-        "attn_k_norm.weight": ("self_attn.k_norm.weight", "attn_norm", "direct", "f32"),
-        "attn_v.weight": ("self_attn.v_proj.weight", "attn_v", "transpose", "q5_k"),
-        "attn_output.weight": ("self_attn.o_proj.weight", "attn_o", "transpose", "q5_k"),
-        "attn_norm.weight": ("input_layernorm.weight", "attn_norm", "direct", "f32"),
-        "ffn_gate_inp.weight": ("mlp.router.weight", "gate", "transpose", "f32"),
-        "exp_probs_b.bias": ("mlp.router.bias", "gate_bias", "direct", "f32"),
-        "ffn_gate_exps.weight": ("mlp.experts.routed.w1", "routed_w1", "routed_expert_transpose", "iq2_xxs"),
-        "ffn_up_exps.weight": ("mlp.experts.routed.w3", "routed_w3", "routed_expert_transpose", "iq2_xxs"),
-        "ffn_down_exps.weight": ("mlp.experts.routed.w2", "routed_w2", "routed_expert_transpose", "iq2_xxs"),
-        "ffn_norm.weight": ("post_attention_layernorm.weight", "ffn_norm", "direct", "f32"),
-    }
+    _GLOBAL_TENSORS = GLOBAL_TENSORS
+    _LAYER_TENSORS = LAYER_TENSORS
 
     def parse_params(self, bundle: GGUFBundle) -> MoEArchitectureParams:
         md = bundle.metadata
@@ -70,27 +52,10 @@ class MiniMaxM2Spec:
         )
 
     def classify_tensor(self, name: str) -> str:
-        if name in self._GLOBAL_TENSORS:
-            return self._GLOBAL_TENSORS[name][1]
-        if not name.startswith("blk."):
-            return "other"
-        suffix = name.split(".", 2)[2] if len(name.split(".", 2)) == 3 else ""
-        return self._LAYER_TENSORS.get(suffix, ("", "other", "", ""))[1]
+        return classify_tensor_name(name)
 
     def build_tensor_mappings(self, bundle: GGUFBundle) -> list[TensorMapping]:
-        params = self.parse_params(bundle)
-        mappings: list[TensorMapping] = []
-        for source_name, (logical, role, transform, _expected_type) in self._GLOBAL_TENSORS.items():
-            mappings.append(TensorMapping(source_name, logical, role, transform))
-        for layer in range(params.n_layers):
-            for suffix, (logical_suffix, role, transform, _expected_type) in self._LAYER_TENSORS.items():
-                source = f"blk.{layer}.{suffix}"
-                logical = f"layers.{layer}.{logical_suffix}"
-                if role in {"routed_w1", "routed_w2", "routed_w3"}:
-                    mappings.append(TensorMapping(source, logical, role, transform, layer=layer))
-                else:
-                    mappings.append(TensorMapping(source, logical, role, transform, layer=layer))
-        return mappings
+        return build_tensor_mappings(self.parse_params(bundle).n_layers)
 
     def _expected_shape_ok(self, source_name: str, dims: tuple[int, ...], params: MoEArchitectureParams) -> bool:
         h = params.hidden_size
