--- src/runtime/moe/__init__.py
@@ -1,1 +1,0 @@
-"""Common MoE execution backends for PocketMoE runtime."""
