+++ tests/bench_cpu_fp4_raw_moe.py
@@ -8,7 +8,7 @@ sys.path.insert(0, str(REPO_ROOT))
 import torch
 
 from src.kernels.ops import Packed4BitWeightAlongK
-from src.runtime.moe.cpu_backend import _load_native_mod
+from src.components.moe.cpu_backend import _load_native_mod
 from src.models.deepseek_v4.runtime import _pack_fp4_weight_rows_for_tile_decode, fp4_block_size
