+++ tests/test_gguf_imatrix.py
@@ -2,13 +2,13 @@ from __future__ import annotations
 
 import numpy as np
 
-from src.gguf.imatrix import _rank_output_path, _write_imatrix_gguf, merge_imatrix_files
-from src.gguf.tensor_reader import GGUFTensorDataReader
+from src.loader.gguf.imatrix import _rank_output_path, _write_imatrix_gguf, merge_imatrix_files
+from src.loader.gguf.tensor_reader import GGUFTensorDataReader
 
 
 def test_layer_capture_helper_respects_collector_filters(monkeypatch) -> None:
-    import src.gguf.imatrix as imatrix
-    from src.runtime.moe.cpu_backend import _layer_imatrix_capture_enabled
+    import src.loader.gguf.imatrix as imatrix
+    from src.components.moe.cpu_backend import _layer_imatrix_capture_enabled
 
     monkeypatch.setenv("DEEPSEEK_IMATRIX_CAPTURE", "1")
     monkeypatch.setenv("DEEPSEEK_IMATRIX_LAYERS", "0")
