+++ tests/test_minimax_m2_moe_runtime.py
@@ -5,9 +5,9 @@ from pathlib import Path
 import pytest
 import torch
 
-from src.gguf.bundle import read_gguf_bundle
+from src.loader.gguf.bundle import read_gguf_bundle
 from src.models.minimax_m2.moe_planning import ROUTED_ROLES, build_minimax_m2_moe_runtime_plan
-from src.runtime.moe.minimax_m2 import (
+from src.models.minimax_m2.moe_runtime import (
     GGUF_DEVICE_TYPE_IDS,
     MiniMaxM2DeviceResidentCache,
     MiniMaxM2RoutedBlockLoader,
